// application/ui/src/features/prompt-sidebar/prompt-modes/visual-prompt/captured-image/zoom/types.ts
/**
 * Copyright (C) 2025 Intel Corporation
 * SPDX-License-Identifier: Apache-2.0
 */

export type Point = {
    x: number;
    y: number;
};

export type Size = { width: number; height: number };

export type ZoomState = {
    scale: number;
    maxZoomIn: number;
    translate: Point;
    hasAnimation: boolean;
    initialCoordinates: Point & { scale: number };
};